==> scripts/02_get_gbif.Rmd <==
---
title: "02_get_gbif"
author: "Dexter H. Locke, PhD"
date: "`r format(Sys.time())`"
output: html_document
editor_options: 
  chunk_output_type: console
---

based off of Millie's "holc-all-spp.R" but designed for laptop

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# 0 set up: load libraries, custom functions, set defaults
```{r}

# load libraries
# packages we'll be using
packs <- c(
    'tidyverse'  # a must have
  , 'gbifdb' # GBIF
  , 'fst' # a faster table, makes outputs files much smaller, too.
  # , 'terra'
  , 'sf'              # spatial support
  , 'mapview'         # webmaps
  , 'janitor'   # cleans things up, also pipe-friendly cross-tabulations
  , 'tictoc'          # times things
  , 'beepr'           # makes noises
)


# check for all of the libraries, install if you don't have them
if (length(setdiff(packs, rownames(installed.packages()))) > 0) {
  install.packages(setdiff(packs, rownames(installed.packages())))  
}

# load them
vapply(packs, library, character.only = TRUE, logical(1), logical.return = TRUE, quietly = TRUE)



# set custom function for getting spatial data
see_sf <- function(){
# what's in memory that are sf - spatial features?
keep(eapply(.GlobalEnv, class),      # gets the objects in the global environment
     ~ any(str_detect(., "sf"))) %>% # selects elements with sf in them
    names(.) %>% as.character(.)     # my simple features
}

see_sf() -> sf_in_memory

# what are the spatial references of those SF classes?
mget(sf_in_memory) %>% purrr::map(~st_crs(.x)$epsg) %>% unlist() #%>% View()


# # get file size
# for(obj in ls()){message(obj); print(object.size(get(obj)), units='auto'); cat('\n')}; rm(obj)

# thanks Phil Donovan @philip_donovan
# https://www.spatialanalytics.co.nz/post/2018/04/01/fixing-st-par/
# Paralise any simple features analysis.
st_parallel <- function(sf_df, sf_func, n_cores, ...){

  # Create a vector to split the data set up by.
  split_vector <- rep(1:n_cores, each = nrow(sf_df) / n_cores, length.out = nrow(sf_df))

  # Perform GIS analysis
  split_results <- split(sf_df, split_vector) %>%
    parallel::mclapply(function(x) sf_func(x, ...), mc.cores = n_cores)


  # Define the output_class. If length is greater than two, then grab the second variable.
  output_class <- class(split_results[[1]])
  if (length(output_class) == 2){
    output_class <- output_class[2]
  }

  # Combine results back together. Method of combining depends on the output from the function.
  if (output_class == "matrix"){
    result <- do.call("rbind", split_results)
    names(result) <- NULL
  } else if (output_class == "sfc") {
    result <- do.call("c", split_results)
    result <- sf_func(result) # do.call combines the list but there are still n_cores of the geometry which had been split up. Running st_union or st_collect gathers them up into one, as is the expected output of these two functions.
  } else if (output_class %in% c('list', 'sgbp') ){
    result <- do.call("c", split_results)
    names(result) <- NULL
  } else if (output_class == "data.frame" ){
    result <- do.call("rbind", split_results)
  } else {
    stop("Unknown class. st_parallel only accepts the following outputs at present: sfc, list, sf, matrix, sgbp.")
  }

  # Return result
  return(result)
}


# custom function for "Not In"
`%nin%` <- Negate(`%in%`)



# fixes mapview
mapviewOptions(fgb = FALSE)


# create directory for large files we don't want on git, one level 
ifelse( !dir.exists('../biodiversity_in_a_concrete_jungle_data_too_big')
       , dir.create('../biodiversity_in_a_concrete_jungle_data_too_big'), FALSE)

# create directory for data to be downloaded, just another organizational level
ifelse( !dir.exists('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads')
       , dir.create('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads'), FALSE)

```

# 1 connect to database
```{r}

db <- gbif_remote()
# # I use a local version of gbif but should work as well with the above code
# # db <- gbif_remote(bucket="gbif", 
# #                   endpoint_override = "minio.cirrus.carlboettiger.info", 
# #                   version="2022-03-01")
# 
```

# 2 sample query - sandbox for working out the queries. Next chunk advances on this, consider this alpha/stale
```{r eval=FALSE, include=FALSE}
# # works
# # grab all occurances in US, count by lat long and class of app
# tic(); US_spp <- db |>
#   filter(countrycode == "US") |>
#   count(class, decimallatitude, decimallongitude) |>
#   collect(); toc(); beep() # ~15 mins
# 
# TIP: load up count with year, collectionCode, institutionCode, basisofrecord then filter?
# 
# # what are the column names available?
# tic(); (test_col_names <- 
#   db |>
#   filter(countrycode == "US") |> 
#   names()); toc(); beep() # lightening fast, nano seconds
# 
# # DOES NOT WORK - don't know the values for stateprovince!
# tic(); (test_download <- 
#   db |>
#   filter(countrycode == "US") |> 
#   filter(stateprovince == "RI") |> 
#   collect()); toc(); beep() # 40 mins and empty

class_to_keep <- c('Aves', 'Insecta', 'Mammalia', 'Reptilia', 'Amphibia')
tic(); (test_download <- 
          db |>
          filter(countrycode == "US") |> 
          select(
              # kingdom - needed?
              class
            , year
            , basisofrecord    # to drop 'unknown'?
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          filter(class %in% class_to_keep) |>    # only classes of interest
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          # combine/reclass data provider here?
          count(  class
                , year
                , basisofrecord
                , collectioncode
                , institutioncode
                , decimallatitude, decimallongitude
                ) |> 
          collect()
        ); toc(); beep() # 36 mins without count, ~9 mins with count
test_download |> distinct(basisofrecord)
test_download |> distinct(collectioncode)
test_download |> distinct(institutioncode)
test_download |> 
  as_tibble() |> 
  write_csv(paste0(getwd()
                   , "/working_data/gbif_download/" # should write to large file directory not for sharing
                   , 'test_'
                   , Sys.Date()
                   , '.csv')
            )
# # try to write out?
# test_download |> 
#   fst::write_fst(
#     paste0(getwd(), "/working_data/gbif_download/test_download.fst"))
# 
# 
# (test_download <- fst::read.fst("test_download.fst") |> tibble())
# 
# 
# (test_download <- fst::read.fst(paste0(getwd(), "/working_data/gbif_download/test_download.fst")) |> tibble())
```

keep_king <- c('Plantae', 'Fungi')
keep_class<- c('Aves', 'Insecta', 'Mammalia', 'Reptilia', 'Amphibia')

## A PLANTS
```{r eval=FALSE, include=FALSE}


tic(); (download_plantae <- 
          db |>
          filter(countrycode == "US") |> 
          select(
              datasetkey
            , kingdom
            , family
            , genus
            , species
            , year
            , basisofrecord    # to drop fossils and material
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          # king/classes of interest
          filter(kingdom == 'Plantae') |> 
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          filter(basisofrecord != 'FOSSIL_SPECIMEN' & basisofrecord != 'MATERIAL_CITATION') |> 
          mutate(source = 
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |> 
          count(
              datasetkey
            , family
            , genus
            , species
            , year
            # , basisofrecord
            # , collectioncode
            # , institutioncode
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
                ) |> 
          collect()
        ); toc(); beep() # ~5 mins

# a few double checks
download_plantae # ~8.5M records
download_plantae |> glimpse()
download_plantae |> tabyl(year)   # growing over time
download_plantae |> tabyl(source) # ~61% from iNaturalist
download_plantae |> tabyl(n_obs) |> 
download_plantae |> arrange(desc(n_obs))

# # write it out ~486 Mb
# download_plantae |>
#   fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/download_plantae_'
#             , Sys.Date()
#             , '.fst'
#             ))
# rm(download_plantae)
```


## B FUNGI
```{r eval=FALSE, include=FALSE}

tic(); (download_fungi <- 
          db |>
          filter(countrycode == "US") |> 
          select(
              datasetkey
            , kingdom
            , family
            , genus
            , species
            , year
            , basisofrecord    # to drop fossils and material
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          # king/classes of interest
          filter(kingdom == 'Fungi') |> 
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          filter(basisofrecord != 'FOSSIL_SPECIMEN' & basisofrecord != 'MATERIAL_CITATION') |> 
          mutate(source = 
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |> 
          count(
              datasetkey
            , family
            , genus
            , species
            , year
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
                ) |> 
          collect()
        ); toc(); beep() # ~5 mins

# a few double checks
download_fungi # ~1M 
download_fungi |> glimpse()
download_fungi |> tabyl(year)
download_fungi |> tabyl(source)
download_fungi |> tabyl(n_obs) |> tibble()
download_fungi |> arrange(desc(n_obs))

# # write it out
# download_fungi |> # ~53Mb
#   fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/download_fungi_'
#             , Sys.Date()
#             , '.fst'
#             ))
# rm(download_fungi)
```


## C AVES - year at a time in a loop
```{r eval=FALSE, include=FALSE}

# create a folder to hold all of the birds data
ifelse( !dir.exists('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/aves')
       , dir.create('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/aves'), FALSE)


# step through each year, downlaod and save directly out. 
for(i in seq(2000, 2020, 1)){
  print(i)
  
  # writing directly out without loading into memory first
  tic(); (db |>
            filter(countrycode == "US") |>
            select(
                datasetkey
              , kingdom
              , class
              , family
              , genus
              , species
              , year
              , basisofrecord    # to drop fossils and material
              , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
              , institutioncode  # has iNaturalist
              , decimallatitude, decimallongitude
              ) |>
            filter(year == i) |> # ONLY A SINGLE YEAR
            # king/classes of interest
            filter(class == 'Aves') |>
            filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
            filter(basisofrecord != 'FOSSIL_SPECIMEN' & basisofrecord != 'MATERIAL_CITATION') |> 
            mutate(source =
                     case_when(
                       str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                       str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                       TRUE ~ 'other')
                   ) |>
            count(
                datasetkey
              , family
              , genus
              , species
              , year
              , source
              , decimallatitude, decimallongitude
              , name = 'n_obs'
                  ) |>
            collect() |>
            fst::write_fst(
              paste0('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/aves/download_aves_'
                                  , i, '_', Sys.Date(), '.fst'))
          ); toc(); beep() 
} #should have timed the whole loop, whoops

# # how long did the years each take?
# tribble(
#   ~year, ~duration,
#     2000, 737.231,
#     2001, 809.129,
#     2002, 965.543,
#     2003, 1007.745,
#     2004, 947.763,
#     2005, 951.937,
#     2006, 938.558,
#     2007, 943.089,
#     2008, 942.358,
#     2009, 939.577,
#     2010, 926.301,
#     2011, 931.678,
#     2012, 938.287,
#     2013, 933.089,
#     2014, 938.959,
#     2015, 956.138,
#     2016, 979.737,
#     2017, 950.34,
#     2018, 1298.071,
#     2019, 2159.562,
#     2020, 2250.91) |> 
#   mutate(minutes = lubridate::seconds(duration) / 60) |> 
#   ggplot(aes(year, minutes)) + 
#   geom_line()

```


## D INSECTA
```{r eval=FALSE, include=FALSE}

tic(); (download_insecta <- 
          db |>
          filter(countrycode == "US") |> 
          select(
              datasetkey
            , kingdom
            , class # needed for insects
            , family
            , genus
            , species
            , year
            , basisofrecord    # to drop fossils and material
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          # king/classes of interest
          filter(class == 'Insecta') |> 
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          filter(basisofrecord != 'FOSSIL_SPECIMEN' & basisofrecord != 'MATERIAL_CITATION') |> 
          mutate(source = 
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |> 
          count(
              datasetkey
            , family
            , genus
            , species
            , year
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
            ) |> 
          collect()
        ); toc(); beep() # ~6 mins

# a few double checks
download_insecta # ~4.2M records
download_insecta |> glimpse()
download_insecta |> tabyl(year) # growth over time..
download_insecta |> tabyl(source) # 70% iNat
download_insecta |> tabyl(n_obs) |> tibble()
download_insecta |> arrange(desc(n_obs))

# # write it out, ~243Mb
# download_insecta |>
#   fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/download_insecta_'
#             , Sys.Date()
#             , '.fst'
#             ))
# rm(download_insecta)
```


## E MAMMALIA
```{r eval=FALSE, include=FALSE}

tic(); (download_mammalia <- 
          db |>
          filter(countrycode == "US") |> 
          select(
              datasetkey
            , kingdom
            , class # needed for mammals
            , family
            , genus
            , species
            , year
            , basisofrecord    # to drop fossils and material
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          # king/classes of interest
          filter(class == 'Mammalia') |> 
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          filter(basisofrecord != 'FOSSIL_SPECIMEN' & basisofrecord != 'MATERIAL_CITATION') |> 
          mutate(source = 
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |> 
          count(
              datasetkey
            , family
            , genus
            , species
            , year
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
                ) |> 
          collect()
        ); toc(); beep() # ~5 mins

# a few double checks
download_mammalia # ~1M records
download_mammalia |> glimpse()
download_mammalia |> tabyl(year)   # not *every* year has more than prior yer!
download_mammalia |> tabyl(source) # ~42% iNat
download_mammalia |> tabyl(n_obs) |> tibble()
download_mammalia |> arrange(desc(n_obs))

# download_mammalia |> 
#   tabyl(year, source) |> 
#   pivot_longer(-year) |> 
#   ggplot(aes(year, value)) + 
#   geom_line(aes(group = name, color = name))


# # write it out, ~40Mb
# download_mammalia |>
#   fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/download_mammalia_'
#             , Sys.Date()
#             , '.fst'
#             ))
# rm(download_mammalia)
```


## F REPTILIA (AKA Squamata!)
```{r eval=FALSE, include=FALSE}


tic(); (download_reptilia <- 
          db |>
          filter(countrycode == "US") |> 
          select(
              datasetkey
            , kingdom
            , class
            , family
            , genus
            , species
            , year
            , basisofrecord    # to drop fossils and material
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          # king/classes of interest
          filter(class == 'Squamata') |> 
          # filter(class == 'Reptilia') |> 
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          filter(basisofrecord != 'FOSSIL_SPECIMEN' & basisofrecord != 'MATERIAL_CITATION') |> 
          mutate(source = 
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |> 
          count(
              datasetkey
            , family
            , genus
            , species
            , year
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
                ) |> 
          collect()
        ); toc(); beep() # ~5 mins

# a few double checks
download_reptilia
download_reptilia |> glimpse()
download_reptilia |> tabyl(year)
download_reptilia |> tabyl(source)
download_reptilia |> tabyl(n_obs)
download_reptilia |> arrange(desc(n_obs))

# # write it out, ~23Mb
# download_reptilia |>
#   fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/download_reptilia_'
#             , Sys.Date()
#             , '.fst'
#             ))
# rm(download_reptilia)
```


## G AMPHIBIA
```{r eval=FALSE, include=FALSE}

tic(); (download_amphibia <- 
          db |>
          filter(countrycode == "US") |> 
          select(
              datasetkey
            , kingdom
            , class
            , family
            , genus
            , species
            , year
            , basisofrecord    # to drop fossils and material
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          # king/classes of interest
          filter(class == 'Amphibia') |> 
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          filter(basisofrecord != 'FOSSIL_SPECIMEN' & basisofrecord != 'MATERIAL_CITATION') |> 
          mutate(source = 
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |> 
          count(
              datasetkey
            , family
            , genus
            , species
            , year
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
                ) |> 
          collect()
        ); toc(); beep() # ~5 mins

# a few double checks
download_amphibia # ~300k records
download_amphibia |> glimpse()
download_amphibia |> tabyl(year)
download_amphibia |> tabyl(source) # 87% iNaturalist
download_amphibia |> tabyl(n_obs)
download_amphibia |> arrange(desc(n_obs))

# # write it out, ~14Mb
# download_amphibia |>
#   fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads/download_amphibia_'
#             , Sys.Date()
#             , '.fst'
#             ))
# rm(download_amphibia)


```

# 3 do the outputs have the same column names?
```{r eval=FALSE, include=FALSE}
# I may not have written out the same files names.. curious so checking

# find the downoaded files, push into tibble 
(col_names <- 
  tibble(filename = list.files('../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads'
                               , recursive = TRUE
                               , full.names = TRUE),
         names = NA))


for(i in seq_along(col_names$filename)){
  print(i)
  # print(col_names$filename[[i]])
  col_names$names[i] <- read_fst(col_names$filename[[i]]) |> colnames() |> tibble()
}

col_names |> 
  unnest_wider(names) |> 
  mutate(name = str_remove(filename
                           , '../biodiversity_in_a_concrete_jungle_data_too_big/gbif_downloads') |>
           str_remove('/aves/download_') |>
           str_remove('/download_') |>
           # str_remove('_2023-01-22.fst')
           str_remove('_2025-02-04.fst')
  ) |> 
  select(name, everything(), -filename) |> 
  View()

# all good!

```


# SANDBOX BELOW

## Test writing directly out without loading the data into memory - for AVES, but using Amphibia as training wheels
```{r eval=FALSE, include=FALSE}
# WORKED GREAT

# try writing directly out without loading into memory first
# Amphibians as example since its small and we know it works
tic(); (db |>
          filter(countrycode == "US") |> 
          select(
              kingdom
            , class
            , genus
            , species
            , year
            # , basisofrecord    # to drop 'unknown'?
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          filter(year >= 2000 & year <= 2020) |> # recent, please
          # king/classes of interest
          filter(class == 'Amphibia') |> 
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          mutate(source = 
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |> 
          count(  #class # TODO I'm not confident a nice kingdom/class look up to genus/species will be available.. keep here?
              genus
            , species
            , year
            # , basisofrecord
            # , collectioncode
            # , institutioncode
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
                ) |> 
          collect() |> 
          fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/download_amphibia_TEST_direct_write_out_'
                        , Sys.Date(), '.fst'))

        ); toc(); beep() # ~6 mins

```

# old defunct example of a single year
### year 2000
```{r eval=FALSE, include=FALSE}

# writing directly out without loading into memory first
tic(); (db |>
          filter(countrycode == "US") |>
          select(
              # kingdom
             class
            , genus
            , species
            , year
            # , basisofrecord    # to drop 'unknown'?
            , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
            , institutioncode  # has iNaturalist
            , decimallatitude, decimallongitude
            ) |>
          # filter(year >= 2000 & year <= 2020) |> # recent, please
          filter(year == 2000) |> # ONLY A SINGLE YEAR
          # king/classes of interest
          filter(class == 'Aves') |>
          filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
          mutate(source =
                   case_when(
                     str_detect(collectioncode, 'EBIRD')  ~  'ebird',
                     str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
                     TRUE ~ 'other')
                 ) |>
          count(  #class # TODO I'm not confident a nice kingdom/class look up to genus/species will be available.. keep here?
              genus
            , species
            , year
            # , basisofrecord
            # , collectioncode
            # , institutioncode
            , source
            , decimallatitude, decimallongitude
            , name = 'n_obs'
                ) |>
          collect() |>
          fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/download_aves_2000_'
                        , Sys.Date(), '.fst'))

        ); toc(); beep() # ~11 mins, ~1M records

```


## AVES - too big? Aaack
```{r eval=FALSE, include=FALSE}
# OLD AND DOES NOT WORK, used looping by year to make data downloads possible.
# # keep_king <- c('Plantae', 'Fungi')
# # keep_class<- c('Aves', 'Insecta', 'Mammalia', 'Reptilia', 'Amphibia')
# 
# tic(); (download_aves_2020 <- 
#           db |>
#           filter(countrycode == "US") |> 
#           select(
#             #   kingdom
#               class
#             , genus
#             , species
#             , year
#             # , basisofrecord    # to drop 'unknown'?
#             , collectioncode   # has GBBC, which is ebird, has EBIRD which is ebird
#             , institutioncode  # has iNaturalist
#             , decimallatitude, decimallongitude
#             ) |>
#           # filter(year >= 2000 & year <= 2020) |> # recent, please
#           filter(year >= 2015 & year <= 2020) |> # recent, please
#           # king/classes of interest
#           filter(class == 'Aves') |> 
#           filter(!is.na(decimallatitude) & !is.na(decimallongitude)) |> # need coords
#           mutate(source = 
#                    case_when(
#                      str_detect(collectioncode, 'EBIRD')  ~  'ebird',
#                      str_detect(institutioncode, 'iNaturalist') ~ 'iNaturalist',
#                      TRUE ~ 'other')
#                  ) |> 
#           count(  #class # TODO I'm not confident a nice kingdom/class look up to genus/species will be available.. keep here?
#               genus
#             , species
#             , year
#             # , basisofrecord
#             # , collectioncode
#             # , institutioncode
#             , source
#             , decimallatitude, decimallongitude
#             , name = 'n_obs'
#                 ) |> 
#           collect()
#         ); toc(); beep() # ~tk mins
# 
# # a few double checks
# download_aves_2015_2020
# download_aves_2015_2020 |> glimpse()
# download_aves_2015_2020 |> tabyl(year) # !!!
# download_aves_2015_2020 |> tabyl(source)
# download_aves_2015_2020 |> tabyl(n)
# 
# # write it out
# download_aves |> fst::write_fst(paste0('../biodiversity_in_a_concrete_jungle_data_too_big/download_aves_2015_2020_'
#             , Sys.Date()
#             , '.csv'
#             ))

```

## End
